+++ AudioPlayer/AudioPlayer/Content/AudioQueue/AudioQueueModel.swift
@@ -12,8 +12,36 @@ class AudioQueueModel {
     @ObservationIgnored
     private var displayLink: DisplayLink?
 
+    var audioTracks: [AudioPlaylist] = []
 
-    init(audioPlayerService: AudioPlayerService) {
+    var currentTrack: AudioTrack?
+
+    init(audioTracksProvider: () -> [AudioPlaylist] = audioQueueTrackProvider, audioPlayerService: AudioPlayerService) {
         self.audioPlayerService = audioPlayerService
+        self.audioTracks = audioTracksProvider()
+    }
+
+    deinit {
+        audioPlayerService.stop()
+    }
+
+    func addNewAudioTrack(url: URL) {
+        let customIndex = audioTracks.firstIndex(where: { $0.id == "Custom" })
+        let audioTrack = AudioTrack(from: .custom(url.absoluteString), status: .idle)
+        let playlist = AudioPlaylist(title: "Custom", tracks: [audioTrack])
+        if let customIndex {
+            let tracks = audioTracks[customIndex].tracks
+            if !tracks.contains(audioTrack) {
+                audioTracks[customIndex].tracks.append(audioTrack)
+            }
+        } else {
+            audioTracks.append(playlist)
+        }
+    }
+
+    func play(_ track: AudioTrack) {
+        if track != currentTrack {
+            currentTrack = track
+        }
     }
 }
